Delete Concrete Cache and make Runtime cache use non-static storage
