+++ docs/v1-to-v2-update.md
@@ -45,3 +45,6 @@ A handler natively supporting APCu was added.
 The `Joomla\Cache\Exception\UnsupportedFormatException` class has been removed. This was previously thrown when an adapter is not
 supported on an environment. Instead, a `Joomla\Cache\CacheItemPoolInterface` interface extending the `Psr\Cache\CacheItemPoolInterface`
 interface has been added and all CacheItemPool implementations offer a static `isSupported()` method to test for support.
+
+### Runtime Adapter
+The Runtime Adapter now stores it's data in a non-static object
